+++ py/sight/configs/question_config.yaml
@@ -1,12 +1,12 @@
-# Q_label1:
-#   desc: "generic question label"
-#   attrs_text_proto: "py/worker/.text_proto_configs/generic.textproto"
+generic:
+  desc: "generic question label"
+  attrs_text_proto: "py/worker/.text_proto_configs/generic.textproto"
 
-# Q_label3:
-#   desc: "FVS question label"
-#   attrs_text_proto: "py/worker/.text_proto_configs/fvs.textproto"
+FVS:
+  desc: "FVS question label"
+  attrs_text_proto: "py/worker/.text_proto_configs/fvs.textproto"
 
-# config attributes for calculator
-calculator:
-  desc: "calculator question label"
-  attrs_text_proto: "py/worker/.text_proto_configs/calculator.textproto"
+# # config attributes for calculator
+# calculator:
+#   desc: "calculator question label"
+#   attrs_text_proto: "py/worker/.text_proto_configs/calculator.textproto"
