+++ py/worker/binary/generic_worker.py
@@ -11,7 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-"""Demo of using the Sight Decision API to run forest simulator."""
+"""Generic worker which propose to worklist_scheduler optimizer."""
 import warnings
 
 
@@ -126,15 +126,6 @@ def main(argv: Sequence[str]) -> None:
     raise app.UsageError("Too many command-line arguments.")
 
   with Sight.create(get_question_label()) as sight:
-
-    # if (FLAGS.parent_id):
-    #   sight_obj = sight_pb2.Object()
-    #   sight_obj.sub_type = sight_pb2.Object.SubType.ST_LINK
-    #   sight_obj.link.linked_sight_id = FLAGS.parent_id
-    #   sight_obj.link.link_type = sight_pb2.Link.LinkType.LT_CHILD_TO_PARENT
-    #   frame = inspect.currentframe().f_back.f_back.f_back
-    #   sight.set_object_code_loc(sight_obj, frame)
-    #   sight.log_object(sight_obj, True)
 
     # this thread checks the outcome for proposed action from server
     decision.init_sight_polling_thread(sight.id,
